requirements.txt
fastapi
pydantic
SQLAlchemy
SQLAlchemy_utils
pyjwt
"passlib[bcrypt]"
